# test: faster test execution in codex-core