--- a/codex-rs/chatgpt/tests/suite/mod.rs
+++ b/codex-rs/chatgpt/tests/suite/mod.rs
@@ -1,0 +1,2 @@
+// Aggregates all former standalone integration tests as modules.
+mod apply_command_e2e;