--- a/codex-rs/exec/tests/suite/mod.rs
+++ b/codex-rs/exec/tests/suite/mod.rs
@@ -1,0 +1,3 @@
+// Aggregates all former standalone integration tests as modules.
+mod apply_patch;
+mod sandbox;