--- a/codex-rs/execpolicy/tests/suite/mod.rs
+++ b/codex-rs/execpolicy/tests/suite/mod.rs
@@ -1,0 +1,10 @@
+// Aggregates all former standalone integration tests as modules.
+mod bad;
+mod cp;
+mod good;
+mod head;
+mod literal;
+mod ls;
+mod parse_sed_command;
+mod pwd;
+mod sed;