--- a/codex-rs/linux-sandbox/tests/suite/mod.rs
+++ b/codex-rs/linux-sandbox/tests/suite/mod.rs
@@ -1,0 +1,2 @@
+// Aggregates all former standalone integration tests as modules.
+mod landlock;