--- a/codex-rs/login/tests/all.rs
+++ b/codex-rs/login/tests/all.rs
@@ -1,0 +1,3 @@
+// Single integration test binary that aggregates all test modules.
+// The submodules live in `tests/suite/`.
+mod suite;