--- a/codex-rs/login/tests/suite/mod.rs
+++ b/codex-rs/login/tests/suite/mod.rs
@@ -1,0 +1,2 @@
+// Aggregates all former standalone integration tests as modules.
+mod login_server_e2e;