--- a/codex-rs/mcp-server/tests/suite/mod.rs
+++ b/codex-rs/mcp-server/tests/suite/mod.rs
@@ -1,0 +1,8 @@
+// Aggregates all former standalone integration tests as modules.
+mod auth;
+mod codex_message_processor_flow;
+mod codex_tool;
+mod create_conversation;
+mod interrupt;
+mod login;
+mod send_message;