--- a/codex-rs/mcp-types/tests/suite/mod.rs
+++ b/codex-rs/mcp-types/tests/suite/mod.rs
@@ -1,0 +1,3 @@
+// Aggregates all former standalone integration tests as modules.
+mod initialize;
+mod progress_notification;