--- a/codex-rs/tui/tests/all.rs
+++ b/codex-rs/tui/tests/all.rs
@@ -1,0 +1,3 @@
+// Single integration test binary that aggregates all test modules.
+// The submodules live in `tests/suite/`.
+mod suite;