… codex-rs/chatgpt/tests/suite/apply_command_e2e.rs
… codex-rs/core/tests/suite/cli_stream.rs
… codex-rs/core/tests/suite/client.rs
… codex-rs/core/tests/suite/compact.rs
… codex-rs/core/tests/suite/exec.rs
… codex-rs/core/tests/suite/exec_stream_events.rs
… codex-rs/core/tests/suite/live_cli.rs
… codex-rs/core/tests/suite/seatbelt.rs
… codex-rs/core/tests/suite/stream_error_allows_next_turn.rs
… codex-rs/core/tests/suite/stream_no_completed.rs
… codex-rs/exec/tests/suite/apply_patch.rs
… codex-rs/exec/tests/suite/sandbox.rs
… codex-rs/execpolicy/tests/suite/bad.rs
… codex-rs/execpolicy/tests/suite/cp.rs
… codex-rs/execpolicy/tests/suite/good.rs
… codex-rs/execpolicy/tests/suite/head.rs
… codex-rs/execpolicy/tests/suite/literal.rs
… codex-rs/execpolicy/tests/suite/ls.rs
… codex-rs/execpolicy/tests/suite/parse_sed_command.rs
… codex-rs/execpolicy/tests/suite/pwd.rs
… codex-rs/execpolicy/tests/suite/sed.rs
… codex-rs/linux-sandbox/tests/suite/landlock.rs
… codex-rs/login/tests/suite/login_server_e2e.rs
… codex-rs/mcp-server/tests/suite/auth.rs
… codex-rs/mcp-server/tests/suite/codex_message_processor_flow.rs
… codex-rs/mcp-server/tests/suite/codex_tool.rs
… codex-rs/mcp-server/tests/suite/create_conversation.rs
… codex-rs/mcp-server/tests/suite/interrupt.rs
… codex-rs/mcp-server/tests/suite/login.rs
… codex-rs/mcp-server/tests/suite/send_message.rs