@bolinfest 2025-08-24
@nornagon-openai thanks for the pointer: it's very helpful!

As per https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html, it would be nice to get rid of these:

```
codex-rs/core/tests/common
codex-rs/mcp-server/tests/common
```

but that can be done in a follow-up PR.

The article also mentions:

> For an internal library, I avoid integration tests all together. Instead, I use Cargo unit tests for “integration” bits:

```
src/
  lib.rs
  tests.rs
  tests/
    foo.rs
    bar.rs
```

>That way, I avoid linking the separate integration tests binary altogether. I also have access to non-pub API of the crate, which is often useful.

I think we should consider doing this for `core` at some point because `core/src/lib.rs` exposes a lot more than I think is healthy, and I think at least some of that was for the benefit of integration tests.

But again, can also be done in a follow-up.